requirements.txt
requests
python-dotenv
langchain-groq
langchain-core
feedparser